LayoutTests/inspector/debugger/pause-in-internal-script-expected.txt

Tests that internal scripts unknown to front-end are processed correctly when appear in debugger call frames. Bug 64995

Debugger was enabled.

Running: testSetBreakpoint
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0)  (pause-in-internal-script.html:14)
    1) forEach ((internal script))
    2) testFunction (pause-in-internal-script.html:12)
Script execution resumed.
Script execution paused.
Debugger was disabled.
